Move holidays test to separate test_holidays module
